More backend refactoring

+++ backend/Services/ProductService.cs
@@ -1,7 +1,6 @@
 ﻿using AutoMapper;
 using backend.Data.Models;
 using backend.DTOs;
-using backend.Models;
 using backend.Repositories;
 
 namespace backend.Services
@@ -11,32 +10,51 @@ namespace backend.Services
         private readonly IProductRepository _productRepository = productRepository;
         private readonly IMapper _mapper = mapper;
 
+        public async Task<IEnumerable<ProductDTO>> GetAllProductsAsync()
+        {
+            IEnumerable<DbProduct> products = await _productRepository.GetAllProductsAsync();
+            return  _mapper.Map<IEnumerable<ProductDTO>>(products);
+        }
+
         public async Task<DbProduct> GetProductByIdAsync(int id)
         {
             return await _productRepository.GetProductByIdAsync(id);
         }
 
-        public async Task<IEnumerable<DbProduct>> GetAllProductsAsync()
+        public async Task<ProductDTO> AddProductAsync(ProductCreationDTO productDTO)
         {
-            return await _productRepository.GetAllProductsAsync();
-        }
-
-        public async Task AddProductAsync(ProductCreationDTO productDTO)
-        {
-            var product = _mapper.Map<ProductCreationDTO>(productDTO);
-            await _productRepository.CreateProductAsync(product);
+            DbProduct dbProduct = new()
+            {
+                Height = productDTO.Height,
+                Width = productDTO.Width,
+                Depth = productDTO.Depth,
+                Weight = productDTO.Weight,
+                Title = productDTO.Title,
+                Description = productDTO.Description,
+                Artist = productDTO.Artist,
+                SellerId = productDTO.SellerId,
+                CreatedAt = DateTime.Now,
+                Year = productDTO.Year
+            };
+            
+            DbProduct product = await _productRepository.CreateProductAsync(dbProduct);
+            return _mapper.Map<ProductDTO>(product);
         }
 
         public async Task UpdateProductAsync(ProductDTO productDTO)
         {
-            var product = _mapper.Map<Product>(productDTO);
+            DbProduct product = await _productRepository.GetProductByIdAsync(productDTO.Id);
+            if (product == null) return;
+
             await _productRepository.UpdateProductAsync(product);
         }
 
         public async Task DeleteProductAsync(int id)
         {
+            DbProduct product = await _productRepository.GetProductByIdAsync(id);
+            if (product == null) return;
+
             await _productRepository.DeleteProductAsync(id);
         }
     }
-
 }
